Simplify keyword replacement (#4)
* Fix bug in looping through images

Caused by accidentally using the dictionary `keys()` method rather than the `items()` method

* Replace custom keyword replacement with docxtpl's custom replacement method

The keyword replacement method is no longer relevant due to the `render()` method accomplishing the same thing. This simplifies the replacement of keywords and also preserves the style of the template document, which was not accomplished with the custom implementation.