Correct bug in retrieving common theme/challenges APG teams
Accidentally was filtering by agency name rather than APG name. In the prior implementation, all APG teams from an agency that had a goal team in with the same theme was included in the output. This was not intended behavior and has been switched to the intended behavior of only linking individual goal teams with the same theme and challenge.